# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sushi API",
    "description": "Документация",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/application": {
      "post": {
        "tags": [
          "view"
        ],
        "summary": "Подать заявку",
        "description": "Подать заявку",
        "operationId": "sendApplication",
        "requestBody": {
          "description": "Апи Токен",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "name, email, contents"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "format": "string",
                    "example": "123"
                  },
                  "email": {
                    "type": "string",
                    "format": "string",
                    "example": "123"
                  },
                  "contents": {
                    "type": "string",
                    "format": "string",
                    "example": "12"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/item/": {
      "post": {
        "tags": [
          "item"
        ],
        "summary": "Поиск товаров",
        "description": "Поиск товаров",
        "operationId": "itemView",
        "requestBody": {
          "description": "Апи Токен",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  ""
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "string",
                    "example": "1"
                  },
                  "category": {
                    "type": "string",
                    "format": "string",
                    "example": "1"
                  },
                  "subcategory": {
                    "type": "string",
                    "format": "string",
                    "example": "1"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Вощврощаются товары",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "user": {
                      "type": "object",
                      "example": []
                    },
                    "token": {
                      "type": "string",
                      "example": "FKOhXAr6Xhx2e6fMdaKZbTOCxCBwLuJDO3j8fYjRoDG9XoAYKQUSPzayU4BM"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  }
}